Move base class to __init__.py
