+++ src/olympia/amo/management/commands/data_load.py
@@ -3,7 +3,7 @@ import os
 from django.core.management import call_command
 from django.core.management.base import CommandError
 
-from ..base import BaseDataCommand
+from .. import BaseDataCommand
 
 
 class Command(BaseDataCommand):
