+++ src/olympia/amo/management/commands/data_seed.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.management import call_command
 
-from ..base import BaseDataCommand
+from .. import BaseDataCommand
 
 
 class Command(BaseDataCommand):
